pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "go-phantom"
version = "0.0.1"
description = "Cross-platform activity simulator"
readme = "README.md"
license = "MIT"
authors = [{name = "Phantom Contributors"}]
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]
dependencies = [
    "pynput>=1.7.6",
    "pyautogui>=0.9.54",
    "pystray>=0.19.4",
    "Pillow>=9.0.0",
    "setproctitle>=1.3.2",
    "rich>=12.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "pyinstaller>=6.0",
]

[project.urls]
Homepage = "https://github.com/hammadxcm/go-phantom"
Repository = "https://github.com/hammadxcm/go-phantom"
Issues = "https://github.com/hammadxcm/go-phantom/issues"

[project.scripts]
phantom = "phantom.__main__:main"

[tool.ruff]
line-length = 99
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]

[tool.setuptools.packages.find]
include = ["phantom*"]
